% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(n=1000,X,Y,plot=TRUE)
}
\arguments{
\item{n}{number of tests}

\item{X}{is the first vector}

\item{Y}{is the second vector}

\item{plot, }{boolean do we want to plot the histogram}
}
\value{
Returns the true correlation, bootstrap correlation, standard error, bias and confidence interval.
}
\description{
Estimate the correlation, by using bootstrap.
}
\examples{
bootstrap(1000,cars$speed,cars$dist,TRUE)
}
\author{
Steffan Leth Jensen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, SDU, Odense \cr
\email{steje14@student.sdu.dk} \cr
}
\keyword{boot}
\keyword{bootstrap}
\keyword{correlation}

