% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffon.R
\name{buffon}
\alias{buffon}
\title{Buffons needle}
\usage{
buffon(n=1000,l=1,d=1)
}
\arguments{
\item{n}{Number of throwed needle}

\item{l}{The length of the needle}

\item{d}{The distance between the lines}
}
\value{
Return the number of hits and the estimated pi.
}
\description{
Estimating pi by throwing needles at lines, whcih is buffons experiment.
}
\examples{
Lazzerini experiment: buffon(3408,2.5,3)
}
\author{
Steffan Leth Jensen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, SDU, Odense \cr
\email{steje14@student.sdu.dk} \cr
}
\keyword{buffon}
\keyword{lazzerini}
\keyword{pi-estimation}

