% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.gof.R
\name{chi.gof}
\alias{chi.gof}
\title{Goodness of fit}
\usage{
chi.gof(x,p)
}
\arguments{
\item{x}{is vector of observed variabel}

\item{p}{the different expected probabilities}
}
\value{
Return the Goodness of fit
}
\description{
Calculate the chi-squared goodness of fit.
}
\examples{
chi.gof(c(2,5),c(0.3,0.7))
}
\author{
Steffan Leth Jensen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, SDU, Odense \cr
\email{steje14@student.sdu.dk} \cr
}
\keyword{chi}
\keyword{fit}
\keyword{goodness}

