% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlot.R
\name{densityPlot}
\alias{densityPlot}
\title{Plot the density}
\usage{
densityPlot(x,n=500, method = "naive" , from, to)
}
\arguments{
\item{x}{A nummeric variable}

\item{n}{Number of plottet points}

\item{method}{Can be either naive or kernel}

\item{from}{Where the function want to start}

\item{to}{Where the function want to end.}
}
\value{
Plots the density function
}
\description{
Plot the density of a method.
}
\examples{
densityPlot(cars$dist,1000,"naive", 0,10)
}
\author{
Steffan Leth Jensen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, SDU, Odense \cr
\email{steje14@student.sdu.dk} \cr
}
\keyword{density}
\keyword{plot}

