% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linMod.R
\name{linMod}
\alias{linMod}
\title{Linear model}
\usage{
linMod(formula)
}
\value{
The function return residual standard error, R-squared, R-squared adjusted, F-statistic, Degrees of freedom, p-value and the betas (this is the intercept with the y-axis and the slope, and so on)
}
\description{
It calculate the linear model. Which can find the regression.
}
\examples{
linMod(cars$speed ~ cars$dist)
}
\author{
Steffan Leth Jensen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, SDU, Odense \cr
\email{steje14@student.sdu.dk} \cr
}
\keyword{linear}
\keyword{model}
\keyword{regretion}
\keyword{residual}

