% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovChain.R
\name{markovChain}
\alias{markovChain}
\title{Markov Chain}
\usage{
markovChain(p,k,n=1000)
}
\arguments{
\item{p}{Is a probability matrix}

\item{k}{The stage where we start}

\item{n}{Number of iteration.}
}
\value{
The stationary probability of each stage
}
\description{
Calculate the probability to be in a stage after n iteration, where we start in stage k.
}
\examples{
markovChain(matrix(c(0.5,0.5,0.2,0.8),2,2,byrow = TRUE),2,1000)
}
\author{
Steffan Leth Jensen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, SDU, Odense \cr
\email{steje14@student.sdu.dk} \cr
}
\keyword{carlo}
\keyword{chain}
\keyword{markov}
\keyword{monte}
\keyword{stage}
\keyword{stationary}

